requirements.txt
torchvision>=0.5.0
pretrainedmodels==0.7.4
efficientnet-pytorch==0.7.1
timm==0.9.5

tqdm
pillow
six

albumentations
pytorch_lightning
matplotlib
pandas
rasterio